refactor(i18n): i18n.py(1627줄)를 keys/legacy_translate/qt_hooks 패키지로 분할
CC-041~046 클린코드 리팩토링. 단일 모듈을 3계층 패키지로 분리하되
__init__.py 재수출 shim으로 기존 12개 import 사이트를 100% 호환 유지.
소비자 파일(main.py, src/ui/*)은 한 줄도 수정하지 않음.

- keys.py: tr() 카탈로그 + 활성 언어 상태(_current_language 단독 소유) + 언어 선택
- legacy_translate.py: 하드코딩 한글 자동 번역(4개 테이블 + 정규식/구/단어/조사 파이프라인)
- qt_hooks.py: PyQt 위젯 텍스트 패치(_qt_i18n_installed 플래그 이동)
- __init__.py: 소비자 이름 16개 순수 재수출 shim

동작 보존 리팩토링:
- CC-044: translate_text를 _apply_regex_pairs / _apply_phrase_substitutions /
  _apply_word_substitutions / _strip_korean_particles 명명 파이프라인으로 분해
- CC-045: 구/단어 테이블 정렬을 매 호출에서 모듈 로드 시 1회 프리컴퓨트(_SORTED_*)로 이동
  (안정 정렬로 동일 길이 키의 tie-order = 삽입순 유지 → 결과 문자열 byte 동일)
- CC-043: 8개 인라인 monkey-patch 블록의 공통 스캐폴딩을 _wrap_callable 제네릭 헬퍼로 통합
  (QToolTip/QMenu/QMessageBox/QFileDialog/헤더라벨 사이트별 인자 semantics는 정확히 보존)
- CC-046/CC-042: Hangul Syllables 블록 경계(가~힣)를 _HANGUL_SYLLABLES_START/END 상수로 통합

tests/test_i18n.py: AST 파싱 경로를 legacy_translate.py로 변경 + found 플래그로 무음 통과 방지,
monkeypatch 대상을 src.core.i18n.keys 모듈로 정정(패키지 attr 패치는 내부 참조에 미적용).

검증: py_compile 통과, tests/test_i18n.py 9/9 통과, 전체 pytest 1809 passed
(macOS validation 1건은 기존 GitHub CI 의존 flaky 테스트로 회귀 판정에서 제외).
